requirements.txt
ccxt==4.2.14
pandas==2.2.0
python-dotenv==1.0.1
numpy>=1.26.0
tqdm>=4.66.0

# API & Web
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
pyyaml>=6.0
python-multipart>=0.0.9

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0
